refactor(dto): (revert) move hydrate to DTO and ensure resolve use hydrate method
