+++ src/Foundation/DataTransferObject.php
@@ -2,14 +2,11 @@
 
 namespace KoalaFacade\DiamondConsole\Foundation;
 
-use CuyZ\Valinor\Mapper\MappingError;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use Illuminate\Support\Traits\Conditionable;
 use Illuminate\Support\Traits\Tappable;
-use KoalaFacade\DiamondConsole\Contracts\DataMapper;
 use KoalaFacade\DiamondConsole\Foundation\DataTransferObject\HasResolvable;
 use Spatie\Cloneable\Cloneable;
 
@@ -91,33 +88,6 @@ abstract readonly class DataTransferObject
         } else {
             $instance = $this->with(...$values);
         }
-
-        return $instance;
-    }
-
-    /**
-     * @param  array<TKey, TValue> | Model  $data
-     *
-     * Hydrate incoming data to resolve unstructured data
-     *
-     * @throws MappingError
-     *
-     * @template TKey of array-key
-     * @template TValue
-     */
-    public static function hydrate(array | Model $data): static
-    {
-        /** @var DataMapper $dataMapper */
-        $dataMapper = resolve(name: DataMapper::class);
-
-        /** @var static $instance */
-        $instance = $dataMapper->execute(
-            signature: static::class,
-            data: match (true) {
-                $data instanceof Model => $data->attributesToArray(),
-                default => $data
-            }
-        );
 
         return $instance;
     }
